max_io_qsize must be large enough to hold NVME_MAX_IO_ENTRIES

+++ lib/libnvmf/libnvmf.h
@@ -63,7 +63,7 @@ struct nvmf_association_params {
 	bool sq_flow_control;		/* SQ flow control required. */
 	bool dynamic_controller_model;	/* Controller only */
 	uint16_t max_admin_qsize;	/* Controller only */
-	uint16_t max_io_qsize;		/* Controller only, 0 for discovery */
+	uint32_t max_io_qsize;		/* Controller only, 0 for discovery */
 	union {
 		struct {
 			uint8_t pda;	/* Tx-side PDA. */
